refactor: isolate video item component
